Install setuptools and pip from apt.
Installing these packages from pip results in CI volatility.  setuptools
60.10.0 appears to interfere with the install paths, unconditionally
inserting `local` in the path so we're left with `$PREFIX/local/bin`
rather than `$PREFIX/bin` as expected.

pip is somewhat of a coattail rider on this commit but is being sourced
from apt for similar reasons.

I've omitted the installation of virtualenv entirely because of parallel
patches currently in flight. That may be revisited before this is ready
for merge.

Co-authored-by: Chris Lalancette <clalancette@openrobotics.org>