Use venv from system instead of pinned virtualenv as venv provider.
virtualenv 16 is broken with Python 3.10 and the pin has been around
as long as it has due to virtualenv detection interfering with our
install-scripts settings for libexec packages.

As part of other patches to enable python3.10 support using a virtualenv
has been moved completely off the table.

If successful we'll have to compare this approach, which may _only_ work
with virtualenv/python/setuptools above a certain threshold but it would
then be on the table.